requirements.txt
Django==1.4
South==0.7.5
psycopg2==2.4.5
django-devserver==0.3.2
django-compressor==1.1.2
django-debug-toolbar==0.9.4
dj-database-url==0.2.1
djangorestframework==2.1.17
-e git://github.com/jgorset/facepy.git@b513145f86082cc24c196cb89e6aa98b17f7912e#egg=facepy
tweepy==1.13
django-filter==0.5.4
PyYAML==3.10

